Library Reference

VirtualEnv.VirtualEnv — Module

VirtualEnv

Module to provide the ability to create self-contained julia virtual environments.

VirtualEnv.Context — Type

Julia instance context definition

VirtualEnv.assets_dir — Method

assets_dir()

Helper function to create assets path

VirtualEnv.check_exists — Method

check_exists(path::String)

Helper function to check if necessary files exist on the system

VirtualEnv.comonicon_build — Function

comonicon_build([sysimg=true]; kwargs...)

Build the CLI manually. This will set system image build to be incremental=true and filter_stdlibs=false to get better compile speed locally. For more detailed reference, please refer to Comonicon documentation.

VirtualEnv.comonicon_install — Method

comonicon_install(;kwargs...)

Install the CLI manually. This will use the default configuration in Comonicon.toml, if it exists. See also comonicon_build. For more detailed reference, please refer to Comonicon documentation.

VirtualEnv.comonicon_install_path — Method

comonicon_install_path([quiet=false])

Install the PATH and FPATH to your shell configuration file. You can use comonicon_install_path(true), to skip interactive prompt. For more detailed reference, please refer to Comonicon documentation.

VirtualEnv.context — Method

context(install_dir::String, depot_dir::String)

Helper function to instantiate a julia context

VirtualEnv.create — Method

create(env_dir::String, clear::Bool, upgrade::Bool, prompt::String)

Helper function to create a single virtual environment

VirtualEnv.sym_or_cp — Method

sym_or_cp(path_orig::String, path_dest::String, sym::Bool)

Helper function that if sym is true, symlink the original path to the destination if sym is false, cp the original path to the destination

VirtualEnv.venv — Function

Creates virtual Julia environments in one or more target directories.

Arguments

env_dirs: One or more directories to create environments in.

Options

-p, --prompt <prompt>: Provides an alternative prompt prefix for this environment. (Default: ENV_DIR)

Flags

-c, --clear: Delete the contents of the environment directory if it already exists. (Default: false)

-u, --upgrade: Upgrade the environment directory to use this version of Julia. (Default: false)